fix: widen byte columns, exclude negative deltas, and stop fabricating dates
compress_original_size and compress_size were int(11), which overflows for
files over 2 GiB; both columns (on sys_file and sys_file_processedfile) are
now bigint(20) unsigned, matching the type TYPO3 core already uses for
sys_file.size.

FileRepository::getTotalBytesSaved() summed compress_original_size minus
compress_size across all compressed rows, so a single compression whose
output grew (still marked compressed) could drag the total negative or
understate real savings; only positive deltas are aggregated now.

CompressionColumnsUpgradeWizard persisted compress_tstamp as time() for
migrated legacy rows, making every one of them appear to have been
compressed on upgrade day. The timestamp is left at its unknown default
instead, and CompressionInfoFormatter::format() now omits the date segment
for a null timestamp rather than defaulting it to now (CompressorTrait
passes the real timestamp explicitly for the live-compression case that
default previously covered).

CompressionInfoFormatter also rendered a malformed double-dash percentage
when the compressed output was larger than the original; the saved-percent
label now switches between a single leading minus and a plain value instead
of always prefixing one.

+++ ext_tables.sql
@@ -1,23 +1,23 @@
 CREATE TABLE sys_file
 (
-    compressed              tinyint(1)   DEFAULT '0' NOT NULL,
+    compressed              tinyint(1)          DEFAULT '0' NOT NULL,
     compress_error          text,
-    compress_provider       varchar(32)  DEFAULT '' NOT NULL,
-    compress_tool           varchar(32)  DEFAULT '' NOT NULL,
-    compress_original_size  int(11)      DEFAULT '0' NOT NULL,
-    compress_size           int(11)      DEFAULT '0' NOT NULL,
-    compress_tstamp         int(11)      DEFAULT '0' NOT NULL,
+    compress_provider       varchar(32)         DEFAULT '' NOT NULL,
+    compress_tool           varchar(32)         DEFAULT '' NOT NULL,
+    compress_original_size  bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_size           bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_tstamp         int(11)             DEFAULT '0' NOT NULL,
     KEY idx_compressed (compressed)
 );
 
 CREATE TABLE sys_file_processedfile
 (
-    compressed              tinyint(1)   DEFAULT '0' NOT NULL,
+    compressed              tinyint(1)          DEFAULT '0' NOT NULL,
     compress_error          text,
-    compress_provider       varchar(32)  DEFAULT '' NOT NULL,
-    compress_tool           varchar(32)  DEFAULT '' NOT NULL,
-    compress_original_size  int(11)      DEFAULT '0' NOT NULL,
-    compress_size           int(11)      DEFAULT '0' NOT NULL,
-    compress_tstamp         int(11)      DEFAULT '0' NOT NULL,
+    compress_provider       varchar(32)         DEFAULT '' NOT NULL,
+    compress_tool           varchar(32)         DEFAULT '' NOT NULL,
+    compress_original_size  bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_size           bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_tstamp         int(11)             DEFAULT '0' NOT NULL,
     KEY idx_compressed (compressed)
 );
